Exec script format code

+++ src/movie/movie.service.ts
@@ -1,5 +1,9 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
-import { Movies } from '@prisma/client';
+import {
+  ConflictException,
+  Injectable,
+  NotFoundException
+} from '@nestjs/common';
+import { Movies, Prisma } from '@prisma/client';
 import { PrismaService } from 'src/prisma.service';
 import { CreateMovieDto } from './dto/create-movie.dto';
 import { UpdateMovieDto } from './dto/update-movie.dto';
@@ -8,20 +12,46 @@ import { UpdateMovieDto } from './dto/update-movie.dto';
 export class MovieService {
   constructor(private readonly prisma: PrismaService) {}
 
+  private readonly _include: Prisma.MoviesInclude = {
+    genders: {
+      select: {
+        name: true
+      }
+    }
+  };
+
   async create(data: CreateMovieDto): Promise<Movies> {
+    console.log(data.genders);
+
+    const movieExists = await this.prisma.movies.findUnique({
+      where: {
+        title_year: {
+          title: data.title,
+          year: data.year
+        }
+      }
+    });
+
+    if (movieExists) throw new ConflictException('Filme já cadastrado!');
+
     const movie = await this.prisma.movies.create({ data });
-    
+
     return movie;
   }
 
   async findAll(): Promise<Movies[]> {
-    return await this.prisma.movies.findMany();
+    return await this.prisma.movies.findMany({
+      include: this._include
+    });
   }
 
   async findOne(id: string): Promise<Movies> {
     const movie = await this.prisma.movies.findUnique({
-      where: { id }
+      where: { id },
+      include: this._include
     });
+
+    if (!movie) throw new NotFoundException('Filme não cadastrado!');
 
     return movie;
   }
@@ -29,10 +59,19 @@ export class MovieService {
   async update(id: string, data: UpdateMovieDto): Promise<Movies> {
     const movieExists = await this.findOne(id);
 
-    if(!movieExists) throw new NotFoundException('Filme não encontrado!');
+    if (!movieExists) throw new NotFoundException('Filme não encontrado!');
 
     const movie = await this.prisma.movies.update({
-      data,
+      data: {
+        ...data,
+        genders: {
+          create: [
+            {
+              name: 'lkska'
+            }
+          ]
+        }
+      },
       where: { id }
     });
 
@@ -40,6 +79,10 @@ export class MovieService {
   }
 
   async remove(id: string): Promise<{ message: string }> {
+    const movieExists = await this.findOne(id);
+
+    if (!movieExists) throw new NotFoundException('Filme não cadastrado!');
+
     await this.prisma.movies.delete({
       where: { id }
     });
